update environment functions to be more conformant and use less memory

+++ src/stdlib/__findenv.c
@@ -2,8 +2,8 @@
 
 extern char **environ;
 
-int __findenv(const char *name) {
-  for (int i = 0; environ[i] != NULL; i++) {
+ssize_t __findenv(const char *name) {
+  for (ssize_t i = 0; environ[i] != NULL; i++) {
     const char *p = strchr(environ[i], '=');
     if (!strncmp(name, environ[i], p - environ[i]))
       return i;
